Version 15.0 Sweet Alert

+++ src/app/pages/api/detail/detail.page.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ConductoresService } from 'src/app/services/api/conductores.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-detail',
@@ -42,8 +43,30 @@ export class DetailPage implements OnInit {
     })
   }
   deleteConductor(){
-    this.apiServices.deleteConductores(this.conductor).subscribe();
-    this.router.navigate(['/apilist']);
+    this.mensaje();
   }
   
+  mensaje(){
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡Esto no se puede deshacer!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, borrarlo',
+      heightAuto: false,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        // Aquí puedes agregar la lógica para el caso afirmativo
+        Swal.fire({
+          icon: 'success',
+          title: 'Conductor Eliminado',
+          heightAuto: false,
+        })
+        this.apiServices.deleteConductores(this.conductor).subscribe();
+        this.router.navigate(['/apilist']);
+      }
+    });
+  }
 }
